feat!: upgrade to Couchbase Lite 4.0.4 (version vectors)
Upgrade the Couchbase Lite dependency from 3.2.4 to 4.0.4 across all
platforms (Java/Android, Objective-C, C) and bump Vector Search to 2.0.0.
This is a breaking release: CBL 4.0 replaces revision trees with version
vectors, changes the default conflict resolution to last-write-wins, and
removes all APIs deprecated in 3.x.

Why: 4.0 aligns mobile with Couchbase Server/Sync Gateway versioning and is
required to keep receiving SDK updates; the version-vector model is one-way
(databases upgrade in place on first open and cannot be reopened by 3.x).

Highlights:
- Deps -> CBL 4.0.4; Vector Search -> 2.0.0 (Maven for JVM/Android, vendored
  C binaries for native, vendored XCFramework for Apple since the 2.0.0 pod
  is not on the CocoaPods trunk)
- ReplicatorConfiguration/CollectionConfiguration reworked to a build-on-
  demand model that preserves the mutable addCollection() API (4.0 makes the
  native config collection-immutable) with same-database/deleted-collection
  validation and default resolution restored across all actuals
- New Document.timestamp and LogDomain.PEER_DISCOVERY/MULTIPEER
- Default ConflictResolver switched to last-write-wins via Document.timestamp
- Removed: legacy logging (Log/ConsoleLogger/FileLogger/Logger/
  LogFileConfiguration -> LogSinks), database-level document/index/listener
  ops (-> Collection), database-based ReplicatorConfiguration members,
  DatabaseConfiguration.mmapEnabled, deprecated FullTextFunction string
  overloads, no-collection Replicator pending APIs, and the deprecated
  database-based config factories
- kermit integration migrated to CustomLogSink; testing-support legacy
  logging plumbing removed
- Tests migrated to the 4.0 API; platform-divergent assertions (CBL
  exception type, getInt overflow) made tolerant; closed-database tests that
  hit an uncatchable Objective-C NSException marked @IgnoreApple
- Regenerated binary-compatibility API dumps; updated CHANGELOG, README,
  roadmap, version -> 4.0.4-2.0.0

Verified: all targets compile; CE JVM 824/824, EE JVM 961/967, CE macOS
815/816; Vector Search verified on JVM and macOS. Remaining failures are
4.0 version-vector / query-planner semantic test reconciliations tracked
as a follow-up.

+++ couchbase-lite/src/appleMain/kotlin/kotbase/logging/FileLogSink.apple.kt
@@ -23,8 +23,8 @@ internal fun CBLFileLogSink.asFileLogSink() = FileLogSink(
     directory,
     usePlaintext,
     maxKeptFiles.toInt(),
-    maxFileSize
+    maxFileSize.toLong()
 )
 
 internal val FileLogSink.actual: CBLFileLogSink
-    get() = CBLFileLogSink(level.actual, directory, isPlainText, maxKeptFiles.toLong(), maxFileSize)
+    get() = CBLFileLogSink(level.actual, directory, isPlainText, maxKeptFiles.toLong(), maxFileSize.toULong())
